+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this library adheres to Rust's notion of
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `subtle::ConstantTimeEq` bound on `ff::Field`
+- `Copy + Send + Sync + 'static` bounds on `ff::PrimeField::Repr`
 
 ## [0.10.1] - 2021-08-11
 ### Added